Update thepiratebay.js
I found that if url is chanfed in line 32 with https://thepiratebay.org it wokrs but only for movian for chrome, because id the user agent if android version is used, piratebay is opened as m.thepiratebay.org and links are different.
User agent should be changed at least on 3 places. I changed it on 2 and i can access magnet url link, but i cant go further because magnet link is from the mobile version. I cant find where to change it.
what i did is to replace 
showtime.httpReq(url).toString();
with
showtime.httpReq(url, {
				                        headers: {
	        	                                        'User-Agent': "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36",
	                	                        },
						}).toString();
help me with the 3rd replace
and is there a way to add search bar.
Thanks.